+++ pesim/math.pyx
@@ -1,10 +1,10 @@
 EPSILON = _EPSILON
 
-cpdef bint feq(float a, float b):
+cpdef bint feq(double a, double b):
     return -_EPSILON < a - b < _EPSILON
 
-cpdef bint flt(float a, float b):
+cpdef bint flt(double a, double b):
     return b - a > _EPSILON
 
-cpdef bint fle(float a, float b):
+cpdef bint fle(double a, double b):
     return b - a > -_EPSILON
